Move event_chain_id_gen out of `DatabasePool`
Turns out `DatabasePool` is also instansiated for the state database,
that doesn't have the `event_auth_chains` table.
